+++ core/src/main/scala/no/nrk/bigquery/Routines.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2020 NRK
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package no.nrk.bigquery
+
+import no.nrk.bigquery.syntax._
+import no.nrk.bigquery.util.{IndexSeqSizedBuilder, Nat, Sized}
+
+object Routines {
+  type Params[N <: Nat] = Sized[IndexedSeq[Param], N]
+  object Params extends IndexSeqSizedBuilder[Param]
+
+  case class Param(name: Ident, maybeType: Option[BQType]) {
+    def definition: BQSqlFrag =
+      maybeType match {
+        case Some(tpe) => bqfr"$name $tpe"
+        case None => bqfr"$name ANY TYPE"
+      }
+  }
+  object Param {
+    def apply(name: String, tpe: BQType): Param =
+      Param(Ident(name), Some(tpe))
+
+    def untyped(name: String): Param =
+      Param(Ident(name), None)
+
+    def fromField(field: BQField): Param =
+      Param(Ident(field.name), Some(BQType.fromField(field)))
+  }
+
+}
