+++ testing/src/test/scala/no/nrk/bigquery/LiveTempUdfTest.scala
@@ -6,18 +6,19 @@
 
 package no.nrk.bigquery
 
+import no.nrk.bigquery.Routines.{Param, Params}
 import no.nrk.bigquery.syntax._
 import no.nrk.bigquery.testing.{BQSmokeTest, BigQueryTestClient}
 
 class LiveTempUdfTest extends BQSmokeTest(BigQueryTestClient.testClient) {
   val udf1 = UDF.temporary(
     Ident("xxdouble_TMP"),
-    UDF.Params(UDF.Param("input", BQType.INT64)),
+    Params(Param("input", BQType.INT64)),
     UDF.Body.Sql(bqsql"(input + input)"),
     Some(BQType.INT64))
   val udf2 = UDF.temporary(
     Ident("half_TMP"),
-    UDF.Params(UDF.Param("input", BQType.INT64)),
+    Params(Param("input", BQType.INT64)),
     UDF.Body.Sql(bqsql"(${udf1(Ident("input"))} / 2)"),
     Some(BQType.FLOAT64))
